Wrap CardsList component in an anonymous function
When using render functions (not methods), any function that should be
local to the render function is usually declared inside of it. However,
every time render is called, the function gets redefined. If we define
it outside of the render function, however, it becomes global to the
current module. The solution is to "simulate" a module by wrapping all
definitions that should be local to the render function in an anonymous
function that will get invoked immediately after its definition. This is
a very well-known pattern in JS.