Open modal when edit card button is clicked
The modal component is a direct child of the board component, even
though it renders with a card as its container. While right now it's not
necessary to do this because CardListCard is a functional component, if
it's ever made into a class, every instance will carry around its own
modal, which is inefective, since only one card can be edited at a time.