Integration (visual) test for popups and modals
Testing if an element is actually visible on screen is not possible with
jsdom since it doesn't do any layout. It has to be visually tested. I
had to go through some hoops in order to make this revision work:

- Speed up animations
- Make click event handling more robust in CardsList/Card
- Assign ids to modal backdrops in order to be able to close the modals
  programmatically
- Fix the intent to disable open/close animations in ActionsMenu

Moreover, since there are now images taken of ActionsMenu, QuickEditCard
and EditCard in the tests for Board, we can delete those and save
ourselves some time!